#1: hoist caps up and remove erronous subcommand usage within flows
