+++ cmd/morc/commands/reqs/reqs.go
@@ -6,15 +6,12 @@ import (
 
 	"github.com/dekarrin/morc"
 	"github.com/dekarrin/morc/cmd/morc/cmdio"
-	"github.com/dekarrin/morc/cmd/morc/commands/reqs/caps"
 	"github.com/dekarrin/morc/cmd/morc/commonflags"
 	"github.com/spf13/cobra"
 )
 
 func init() {
 	RootCmd.PersistentFlags().StringVarP(&commonflags.ProjectFile, "project_file", "F", morc.DefaultProjectPath, "Use the specified file for project data instead of "+morc.DefaultProjectPath)
-
-	RootCmd.AddCommand(caps.CapsCmd)
 }
 
 var RootCmd = &cobra.Command{
